PipeConnectionBase
